stype: refactoring
