+++ src/components/Button/Button.stories.tsx
@@ -3,8 +3,8 @@ import { storiesOf } from '@storybook/react'
 import * as React from 'react'
 import styled from 'styled-components'
 import { PrimaryButton, PrimaryButtonAnchor } from './PrimaryButton'
-import { DangerButton, DangerButtonAnchor } from './DangerButton'
 import { SecondaryButton, SecondaryButtonAnchor } from './SecondaryButton'
+import { DangerButton, DangerButtonAnchor } from './DangerButton'
 import { SkeletonButton, SkeletonButtonAnchor } from './SkeletonButton'
 import { Icon } from '../Icon'
 import readme from './README.md'
