+++ src/components/Button/SkeletonButton.tsx
@@ -1,31 +1,17 @@
 import * as React from 'react'
 import styled, { css } from 'styled-components'
 
-import { buttonFactory, BaseProps } from './BaseButton'
+import {
+  BaseButton,
+  BaseButtonAnchor,
+  MargedButtonProps,
+  MargedAnchorProps,
+  ButtonProps,
+  AnchorProps,
+} from './BaseButton'
 
 import { InjectedProps, withTheme } from '../../hocs/withTheme'
 import { isTouchDevice } from '../../libs/ua'
-
-interface ClickEvent {
-  preventDefault: () => void
-}
-
-interface ButtonProps extends BaseProps {
-  onClick?: (e: ClickEvent) => void
-  disabled?: boolean
-}
-
-interface AnchorProps extends BaseProps {
-  href: string
-  target?: string
-  rel?: string
-}
-
-type MargedButtonProps = ButtonProps & InjectedProps
-type MargedAnchorProps = AnchorProps & InjectedProps
-
-const ButtonComponent: React.FC<MargedButtonProps> = buttonFactory<ButtonProps>('button')
-const ButtonAnchorComponent: React.FC<MargedAnchorProps> = buttonFactory<AnchorProps>('a')
 
 const injectStyle = (
   component: React.FC<MargedButtonProps> | React.FC<MargedAnchorProps>,
@@ -51,7 +37,5 @@ const injectStyle = (
   }}
 `
 
-export const SkeletonButton: React.FC<ButtonProps> = withTheme(injectStyle(ButtonComponent))
-export const SkeletonButtonAnchor: React.FC<AnchorProps> = withTheme(
-  injectStyle(ButtonAnchorComponent),
-)
+export const SkeletonButton: React.FC<ButtonProps> = withTheme(injectStyle(BaseButton))
+export const SkeletonButtonAnchor: React.FC<AnchorProps> = withTheme(injectStyle(BaseButtonAnchor))
